models/user.py
from klapi import db
from sqlalchemy.orm import relationship

class User(db.Model):
    __tablename__ = 'user'
    
    id = db.Column(db.Integer, primary_key=True)
    username = db.Column(db.String(50), unique=True, nullable=False)
    password = db.Column(db.String(200), nullable=False)
    role = db.Column(db.String(50), nullable=False)
    
    reservas = db.relationship('Reserva', back_populates='cliente')