requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
gunicorn
